# Conflicts:
#	pom.xml

+++ src/main/java/uz/pdp/kiyim_online_dokon/entity/Transactions.java
@@ -16,18 +16,17 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "transactions") // Jadval nomi 'transactions' deb belgilandi
+@Table(name = "transactions")
 public class Transactions {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // Integer o'rniga JPA standarti bo'yicha Long
+    private Integer id;
 
-    @OneToOne // Order bilan One-to-One bog'liqlik
+    @OneToOne
     @JoinColumn(name = "order_id", unique = true, nullable = false)
     private Orders order;
 
-    // UserID o'rniga Users obyekti
-    // Tranzaksiya Orders obyekti orqali foydalanuvchiga bog'langan bo'ladi, lekin uni to'g'ridan-to'g'ri bog'lash ham mumkin:
+
     @ManyToOne
     @JoinColumn(name = "user_id")
     private Users user;
@@ -43,16 +42,15 @@ public class Transactions {
     private PaymentMethod paymentMethod;
 
     @Column(unique = true)
-    private String providerTransactionId; // payment_providerdan kelgan tranzaksiya IDsi
+    private String providerTransactionId;
 
     @CreationTimestamp
     private LocalDateTime createdAt;
 
-    // Yangilanish vaqti uchun:
-    // @UpdateTimestamp kabi anotatsiya Hibernate dan foydalanishda qo'shilishi mumkin.
+
     private LocalDateTime updatedAt;
 
-    // Eslatma: `row_response` (xom javob) kabi texnik maydonlar ba'zan oddiy String sifatida saqlanadi.
+
     @Column(columnDefinition = "TEXT")
     private String rawResponse;
 }
